Added the IamClientRequest class and created a Impl file for the IamService class
